Event options moved to TornadoOptions
